# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResearchInput {
  /**
     * The market research query, e.g. 'Analyze Tesla solar business' or 'Coffee shop market in Lahore'
     * @minLength 3
     */
  query: string;
}

export type ResearchReportStatus = typeof ResearchReportStatus[keyof typeof ResearchReportStatus];


export const ResearchReportStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface SwotAnalysis {
  strengths?: string[];
  weaknesses?: string[];
  opportunities?: string[];
  threats?: string[];
}

export interface Competitor {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  marketShare?: string | null;
  /** @nullable */
  pricing?: string | null;
  /** @nullable */
  strengths?: string | null;
}

export interface ResearchReport {
  id: number;
  query: string;
  status: ResearchReportStatus;
  /**
     * Which agent is currently running
     * @nullable
     */
  currentAgent?: string | null;
  /**
     * Short description of current progress
     * @nullable
     */
  agentProgress?: string | null;
  /**
     * Full markdown report (populated when status=completed)
     * @nullable
     */
  report?: string | null;
  swotAnalysis?: SwotAnalysis;
  /** @nullable */
  competitors?: Competitor[] | null;
  /** @nullable */
  marketSize?: string | null;
  /** @nullable */
  keyTrends?: string[] | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type ResearchStatusUpdateStatus = typeof ResearchStatusUpdateStatus[keyof typeof ResearchStatusUpdateStatus];


export const ResearchStatusUpdateStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ResearchStatusUpdate {
  id: number;
  status: ResearchStatusUpdateStatus;
  /** @nullable */
  currentAgent?: string | null;
  /** @nullable */
  agentProgress?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface ResearchStats {
  total: number;
  completed: number;
  running: number;
  failed: number;
  recentQueries?: string[];
}