# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Competitor {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  marketShare?: string | null;
  /** @nullable */
  pricing?: string | null;
  /** @nullable */
  strengths?: string | null;
}