# lib/api-zod/src/generated/types/researchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchInput {
  /**
     * The market research query, e.g. 'Analyze Tesla solar business' or 'Coffee shop market in Lahore'
     * @minLength 3
     */
  query: string;
}