# lib/api-zod/src/generated/types/researchReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Competitor } from './competitor';
import type { ResearchReportStatus } from './researchReportStatus';
import type { SwotAnalysis } from './swotAnalysis';

export interface ResearchReport {
  id: number;
  query: string;
  status: ResearchReportStatus;
  /**
     * Which agent is currently running
     * @nullable
     */
  currentAgent?: string | null;
  /**
     * Short description of current progress
     * @nullable
     */
  agentProgress?: string | null;
  /**
     * Full markdown report (populated when status=completed)
     * @nullable
     */
  report?: string | null;
  swotAnalysis?: SwotAnalysis;
  /** @nullable */
  competitors?: Competitor[] | null;
  /** @nullable */
  marketSize?: string | null;
  /** @nullable */
  keyTrends?: string[] | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}