# lib/api-zod/src/generated/types/researchStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchStatusUpdateStatus } from './researchStatusUpdateStatus';

export interface ResearchStatusUpdate {
  id: number;
  status: ResearchStatusUpdateStatus;
  /** @nullable */
  currentAgent?: string | null;
  /** @nullable */
  agentProgress?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}