# lib/api-zod/src/generated/types/swotAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Market Research & Competitor Analyst Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SwotAnalysis {
  strengths?: string[];
  weaknesses?: string[];
  opportunities?: string[];
  threats?: string[];
}